# contracts/openapi/abfall-api.yaml
openapi: 3.1.0
info:
  title: Abfall APP API
  version: 0.2.0
  description: Contract-first API für den synthetischen, lokal testbaren Phase-1/2-Piloten.
  license:
    name: UNLICENSED
    identifier: LicenseRef-Proprietary
servers:
  - url: /
    description: Aktueller API-Ursprung; konkrete Basis-URLs kommen aus der Laufzeitkonfiguration.
security: []
...
paths:
...
  /v1/tenants:
    get:
      operationId: getMunicipalities
      summary: Listet alle öffentlich auswählbaren teilnehmenden Kommunen.
      tags: [Tenants]
      responses:
        "200":
          description: Alphabetische Kommunenliste.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MunicipalitySummary"
        "503":
          description: Kommunenverzeichnis ist vorübergehend nicht verfügbar.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "429":
          description: Zu viele Anfragen.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
  /v1/tenants/{tenantKey}/config:
    get:
      operationId: getTenantConfig
      summary: Liefert die öffentliche Laufzeitkonfiguration eines Mandanten.
      tags:
        - Tenants
      parameters:
        - name: tenantKey
          in: path
          required: true
          description: Stabiler, öffentlicher Mandantenschlüssel.
          schema:
            type: string
            pattern: "^[a-z0-9][a-z0-9-]{0,62}$"
      responses:
        "200":
          description: Öffentliche Mandantenkonfiguration.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantConfig"
              examples:
                demo:
                  $ref: "../examples/demo-tenant-config.json"
        "404":
          description: Der Mandant ist nicht konfiguriert.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
...
components:
  securitySchemes:
    monitoringToken:
      type: apiKey
      in: header
      name: X-Monitoring-Token
      description: Separates, mindestens 32 Zeichen langes Geheimnis für den Qualitätsagenten.
...
  schemas:
...
    Branding:
      type: object
      additionalProperties: false
      required:
        - logoUrl
        - primaryColor
        - infoColor
      properties:
        logoUrl:
          type: string
        primaryColor:
          type: string
          pattern: "^#[0-9A-Fa-f]{6}$"
        infoColor:
          type: string
          pattern: "^#[0-9A-Fa-f]{6}$"
    LegalLinks:
      type: object
      additionalProperties: false
      required:
        - imprint
        - privacy
        - accessibility
      properties:
        imprint:
          type: string
        privacy:
          type: string
        accessibility:
          type: string
    SupportContact:
      type: object
      additionalProperties: false
      required:
        - label
        - email
        - phone
      properties:
        label:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
    ServiceArea:
      type: object
      additionalProperties: false
      required: [city, reportingOffice, phone, email]
      properties:
        city:
          type: string
        reportingOffice:
          type: string
        phone:
          type: string
        email:
          type: string
          format: email
    TenantConfig:
      type: object
      additionalProperties: false
      required:
        - tenantId
        - name
        - shortName
        - branding
        - timezone
        - locales
        - enabledFeatures
        - legalLinks
        - serviceArea
        - supportContacts
        - contentVersion
      properties:
        tenantId:
          type: string
        name:
          type: string
        shortName:
          type: string
        branding:
          $ref: "#/components/schemas/Branding"
        timezone:
          type: string
        locales:
          type: array
          minItems: 1
          items:
            type: string
        enabledFeatures:
          type: object
          additionalProperties:
            type: boolean
        legalLinks:
          $ref: "#/components/schemas/LegalLinks"
        serviceArea:
          $ref: "#/components/schemas/ServiceArea"
        supportContacts:
          type: array
          items:
            $ref: "#/components/schemas/SupportContact"
        contentVersion:
          type: string
...
    MunicipalitySummary:
      type: object
      additionalProperties: false
      required: [tenantId, name, shortName, city, primaryColor]
      properties:
        tenantId: { type: string }
        name: { type: string }
        shortName: { type: string }
        city: { type: string }
        primaryColor: { type: string, pattern: "^#[0-9A-Fa-f]{6}$" }
...
    Problem:
      type: object
      additionalProperties: true
      required:
        - type
        - title
        - status
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
        instance:
          type: string
          format: uri-reference